PSLogHandler multi-line analysis & regex tests/fixes
AnalyzeLine can now request additional lines by returning the input or return null to continue with a new analysis.
Any additional lines that were not relevant to the previous line can be added to the reAnalysisQueue.
Anything in the queue will be analysed before any new lines are read from the log file. It has no impact on whether or not additional lines are requested for the same analysis.
This queue is passed as a variable since it can get messy due to multithreading & multiple file readers otherwise. Specifying ref to clarify.