(Mostly) Fix flicker on window resize by resizing off screen
For future reference:
Still not completely the way I'd like it but it's an improvement.
After trying both, it seems that resizing occurs before movement.
So table needs to be moved off-screen with it's current size and call for resize when moving to it's final position.
Flicker now rarely occurs but 200ms doesnt seem to be enough time to draw the entire thing correctly. Play with it a little.